crowd_source/crowd_server/requirements/local.txt
-r tests.txt

django-debug-toolbar==3.6.0
ipython==8.4.0
gunicorn==20.0.4
django-rest-knox==4.2.0
whitenoise==6.2.0